# Add benchmarking project and remove empty array allocations