--- a/src/CsvHelper.Benchmarks/BenchmarkMain.cs
+++ b/src/CsvHelper.Benchmarks/BenchmarkMain.cs
@@ -1,0 +1,11 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace CsvHelper.Benchmarks;
+
+internal class BenchmarkMain
+{
+	static void Main(string[] args)
+	{
+		_ = BenchmarkRunner.Run<BenchmarkEnumerateRecords>();
+	}
+}