--- a/src/CsvHelper.Benchmarks/CsvHelper.Benchmarks.csproj
+++ b/src/CsvHelper.Benchmarks/CsvHelper.Benchmarks.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CsvHelper\CsvHelper.csproj" />
+  </ItemGroup>
+
+</Project>