--- a/src/CsvHelper/ObjectCreator.cs
+++ b/src/CsvHelper/ObjectCreator.cs
@@ -53,6 +53,11 @@ public class ObjectCreator
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	private static Type[] GetArgTypes(object?[] args)
 	{
+		if (args.Length == 0)
+		{
+			return Array.Empty<Type>();
+		}
+
 		var argTypes = new Type[args.Length];
 		for (var i = 0; i < args.Length; i++)
 		{