- This PR adds a CsvHelper.Benchmarks project to hold onto benchmarks for CsvHelper. This also adds a BenchmarkEnumerateRecords benchmark that benchmarks iterating through a CSV using csv.GetRecords()

- This PR updates ObjectCreator.GetArgTypes to not allocate an empty Type[] when the passed in object array is empty. This helps reduce allocations by ~14% in the new BenchmarkEnumerateRecords benchmark.

## Before
| Method           | Mean     | Error     | StdDev    | Gen0     | Gen1    | Allocated |
|----------------- |---------:|----------:|----------:|---------:|--------:|----------:|
| EnumerateRecords | 2.951 ms | 0.0576 ms | 0.0640 ms | 203.1250 | 15.6250 |   1.69 MB |

![image](https://github.com/user-attachments/assets/34e52e76-f87d-4278-a002-d40019972707)

## After
| Method           | Mean     | Error     | StdDev    | Gen0     | Gen1    | Allocated |
|----------------- |---------:|----------:|----------:|---------:|--------:|----------:|
| EnumerateRecords | 2.303 ms | 0.0171 ms | 0.0152 ms | 171.8750 | 15.6250 |   1.46 MB |

![image](https://github.com/user-attachments/assets/07d4c86e-3c31-4813-b94b-2048ea925617)